Moved RulesAmendmentsRaw to DTO folder and renamed to RulesAmendmentsDTO
