+++ src/main/kotlin/nomic/data/repositories/rulesamendments/IRuleAmendmentRepository.kt
@@ -1,11 +1,11 @@
 package nomic.data.repositories.rulesamendments
 
-import nomic.domain.entities.RulesAmendmentsRaw
+import nomic.data.dtos.RulesAmendmentsDTO
 
 /**
  * Interface returns the obtained query when prompting the DB for Rules and Amendments
  *
- * @see [nomic.domain.entities.RulesAmendmentsRaw]
+ * @see [nomic.data.dtos.RulesAmendmentsDTO]
  */
 interface IRuleAmendmentRepository {
     /**
@@ -14,5 +14,5 @@ interface IRuleAmendmentRepository {
      * @param gameId The id of the game to collect rule and amendment data on
      * @return Raw rule and amendment data for a specific game
      */
-    fun getRulesAmendments(gameId: Int): MutableList<RulesAmendmentsRaw>
+    fun getRulesAmendments(gameId: Int): MutableList<RulesAmendmentsDTO>
 }
